pyproject.toml
[project]
name = "smart-auto-editor"
version = "0.1.0"
description = "smart editor automation using python"
authors = [
    {name = "vinisvictorelli",email = "vini.svictorelli@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
dependencies = [
    "torch (==2.2.2)",
    "numpy (<2)",
    "moviepy (==2.1.1)",
    "pydub (==0.25.1)",
    "speechrecognition (==3.11.0)",
    "streamlit (==1.40.2)",
    "openai-whisper (>=20240930,<20240931)",
    "certifi (>=2024.12.14,<2025.0.0)",
    "textual (>=1.0.0,<2.0.0)"
]
[tool.poetry.group.dev.dependencies]
textual = {extras = ["dev"], version = ">=0.11.0"}


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
package-mode = false
build-backend = "poetry.core.masonry.api"
